+++ backend/canisters/user_index/api/src/updates/award_external_achievement.rs
@@ -1,0 +1,24 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::UserId;
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub user_id: UserId,
+    pub name: String,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(SuccessResult),
+    InvalidCaller,
+    NotFound,
+    InsufficientBudget,
+    Expired,
+    AlreadyAwarded,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SuccessResult {
+    pub remaining_chit_budget: u32,
+}