+++ frontend/openchat-agent/src/services/openchatAgent.ts
@@ -1654,11 +1654,6 @@ export class OpenChatAgent extends EventTarget {
             groupsAdded = userResponse.groupChats.summaries;
             communitiesAdded = userResponse.communities.summaries;
 
-            if (userResponse.groupChats.cached !== undefined) {
-                currentGroups = userResponse.groupChats.cached.summaries;
-                latestActiveGroupsCheck = userResponse.groupChats.cached.timestamp;
-            }
-
             avatarId = userResponse.avatarId;
             blockedUsers = userResponse.blockedUsers;
             pinnedGroupChats = userResponse.groupChats.pinned;
@@ -1794,6 +1789,22 @@ export class OpenChatAgent extends EventTarget {
             groupIndexResponse.deletedCommunities.forEach((c) => groupsRemoved.add(c.id));
 
             latestActiveGroupsCheck = groupIndexResponse.timestamp;
+
+            // Also check for updates for recently joined groups and communities since it may take a few iterations
+            // before the GroupIndex knows that they are active
+            const recentlyJoinedCutOff = BigInt(start - 10 * 60 * 1000);
+
+            for (const group of currentGroups) {
+                if (group.membership.joined > recentlyJoinedCutOff) {
+                    groupsToCheckForUpdates.add(group.id.groupId);
+                }
+            }
+
+            for (const community of currentCommunities) {
+                if (community.membership.joined > recentlyJoinedCutOff) {
+                    groupsToCheckForUpdates.add(community.id.communityId);
+                }
+            }
         }
 
         const byLocalUserIndex: Map<string, GroupAndCommunitySummaryUpdatesArgs[]> = new Map();